Clean up IncomingHandler
Move Message to messages.rs as it makes more sense there
Add some more documentation and comments.
Solves #9
